+++ platformio.ini
@@ -7,12 +7,13 @@
 ;
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
-[platformio]
-default_envs = lineman_smallmot
+; [platformio]
+; default_envs = lineman_smallmot
 
 [env]
 monitor_speed = 115200
 ; upload_speed = 115200
+; upload_speed = 2000000
 platform = espressif32
 framework = arduino
 board = nodemcu-32s
@@ -28,16 +29,16 @@ build_src_filter =
 	+<Drive/Drive.cpp>
 	+<Drive/Drive.h>
 	+<Robot/MotorControl.h>
-  +<Robot/MotorControl.cpp>
-	+<Robot/Lights.h>
-  +<pairing.h>
+    +<Robot/MotorControl.cpp>
+    ; +<Robot/Lights.h>
+    +<pairing.h>
 	+<pairing.cpp>
-  +<builtInLED.h>
-  +<builtInLED.cpp>	
-  -<depairingStation.cpp>
-  -<Drive/DriveQuick.cpp>
-  -<Drive/DriveQuick.h>
-  -<Robot/Quarterback_old.cpp>
+    +<builtInLED.h>
+    +<builtInLED.cpp>	
+    ; -<depairingStation.cpp>
+    ; -<Drive/DriveQuick.cpp>
+    ; -<Drive/DriveQuick.h>
+    ; -<Robot/Quarterback_old.cpp>
 
 ; bot numbering:
 ; 0 = Lineman
@@ -55,68 +56,52 @@ build_src_filter =
 build_flags = 
 	-D BOT_TYPE=0
 	-D MOTOR_TYPE=1
-; build_src_filter = 
-	; +<main.cpp> 
 
 [env:lineman_bigmot]
 build_flags = 
     -D BOT_TYPE=0
     -D MOTOR_TYPE=0
-; build_src_filter = 
-	; +<main.cpp>
 
 [env:receiver]
-build_flags = -D BOT_TYPE=1
-; build_src_filter = 
-	; +<Drive/Drive.cpp>
-	; +<Drive/Drive.h>
-	; ; Work in progress
-	; +<PolarRobotics.h>
-	; +<main_special.cpp>
+build_flags = 
+    -D BOT_TYPE=1
 
 [env:center]
-build_flags = -D BOT_TYPE=2
+build_flags = 
+    -D BOT_TYPE=2
 build_src_filter = 
-	; +<Drive/Drive.cpp>
-	; +<Drive/Drive.h>
+	${env.build_src_filter} 
 	+<Robot/Center.h>
-	+<Robot/Center.cpp>
-	; +<PolarRobotics.h>
-	; +<main_special.cpp>
 
 [env:mecanum_center]
-build_flags = -D BOT_TYPE=3
+build_flags = 
+    -D BOT_TYPE=3
 
 
 [env:quarterback]
 build_flags = -D BOT_TYPE=4
 build_src_filter = 
-	; +<Drive/Drive.cpp>
-	; +<Drive/Drive.h>
+    ${env.build_src_filter} 
 	+<Robot/Quarterback.h>
-	; +<PolarRobotics.h>
-	; +<main_special.cpp>
 
 [env:kicker]
 build_flags = -D BOT_TYPE=5
 build_src_filter = 
-	; +<Drive/Drive.cpp>
-	; +<Drive/Drive.h>
-	+<Robot/kicker.h>
-	; +<PolarRobotics.h>
-	; +<main_special.cpp>
+    ${env.build_src_filter} 
+	+<Robot/Kicker.h>
 
 [env:depairing]
 build_src_filter = 
-  -<*>
-  +<depairingStation.cpp>
-  +<pairing.h>
-  +<pairing.cpp>
-  +<PolarRobotics.h>
-  +<builtInLED.h>
-  +<builtInLED.cpp>
+    -<*>
+    +<depairingStation.cpp>
+    +<pairing.h>
+    +<pairing.cpp>
+    +<PolarRobotics.h>
+    +<builtInLED.h>
+    +<builtInLED.cpp>
 
 [env:address]
 build_src_filter = 
 	-<*>
 	+<getAddress.cpp>
+
