+++ src/PolarRobotics.h
@@ -8,19 +8,10 @@
 #define LEFT_MOT_PIN  32  // Drive Base
 #define RIGHT_MOT_PIN 33
 
-#define SPECBOT_PIN1  0   // Special Bot Pins
-#define SPECBOT_PIN2  0
-#define SPECBOT_PIN3  0
-#define SPECBOT_PIN4  0
-
-#define CENT_ARM_PIN  0   // Old Center
-#define CENT_CLAW_PIN 0
-
-#define QB_FLYWH_PIN  0   // Quarterback
-#define QB_CONV_PIN   0
-#define QB_ELVMOT_PIN 0
-
-#define KICK_WIND_PIN 0   // Kicker
+#define SPECBOT_PIN1  18   // Special Bot Pins
+#define SPECBOT_PIN2  19
+#define SPECBOT_PIN3  4
+#define SPECBOT_PIN4  2
 
 // Robot Type Enum
 // 0 for lineman, 1 for reciever, 2 for center, 3 for quarterback, 4 for kicker
